+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/utils/TelegramAPIUrlsKeeper.kt
@@ -50,7 +50,7 @@ class TelegramAPIUrlsKeeper(
      * @param rawData Data from [dev.inmo.tgbotapi.webapps.WebApp.initData]
      * @param hash Data from [dev.inmo.tgbotapi.webapps.WebApp.initDataUnsafe] from the field [dev.inmo.tgbotapi.webapps.WebAppInitData.hash]
      */
-    fun checkWebAppLink(rawData: String, hash: String): Boolean {
+    fun checkWebAppData(rawData: String, hash: String): Boolean {
         val preparedData = rawData
             .decodeURLQueryComponent()
             .split("&")
@@ -60,4 +60,11 @@ class TelegramAPIUrlsKeeper(
 
         return HMAC.hmacSHA256(webAppDataSecretKeyHash.bytes, preparedData.toByteArray()).hexLower == hash.lowercase()
     }
+
+    /**
+     * @param rawData Data from [dev.inmo.tgbotapi.webapps.WebApp.initData]
+     * @param hash Data from [dev.inmo.tgbotapi.webapps.WebApp.initDataUnsafe] from the field [dev.inmo.tgbotapi.webapps.WebAppInitData.hash]
+     */
+    @Deprecated("Renamed", ReplaceWith("checkWebAppData"))
+    inline fun checkWebAppLink(rawData: String, hash: String) = checkWebAppData(rawData, hash)
 }